+++ src/storage.rs
@@ -41,7 +41,7 @@ use crate::util::Uppercase;
 use chrono::DateTime;
 use smol_str::SmolStr;
 use sql::ir::operator::ConflictStrategy;
-use sql::ir::options::ReadPreference;
+use sql::ir::options::{Forward, ReadPreference};
 use std::cell::RefCell;
 use std::collections::hash_map::Entry;
 use std::collections::{HashMap, HashSet};
@@ -3367,16 +3367,24 @@ impl DbConfig {
         ReadPreference::try_from(raw).expect("invalid read_preference value")
     }
 
+    #[inline]
+    pub fn forward(&self) -> Forward {
+        let raw = config::DYNAMIC_CONFIG.read_preference.current_value();
+        Forward::try_from(raw).expect("invalid forward value")
+    }
+
     /// Gets sql options as [`sql::ir::options::Options`] struct.
     #[inline]
     pub fn sql_query_options(&self) -> sql::ir::options::Options {
         let sql_vdbe_opcode_max = self.sql_vdbe_opcode_max();
         let sql_motion_row_max = self.sql_motion_row_max();
         let read_preference = self.read_preference();
+        let forward = self.forward();
         sql::ir::options::Options {
             sql_motion_row_max,
             sql_vdbe_opcode_max,
             read_preference,
+            forward,
             ..Default::default()
         }
     }