+++ test/int/sql/explain.sql
@@ -1146,7 +1146,7 @@ BEGIN
   UPDATE t SET a = a * 2 WHERE pk = 3;
 END $$;
 -- ERROR:
-sbroad: LOGICAL and BUCKETS modes for explain are not implemented for transactions
+sbroad: LOGICAL, BUCKETS, and FORWARD modes for explain are not implemented for transactions
 
 -- TEST: explain-query-different-buckets
 -- SQL: